requirements.txt
# Azure AI and Core SDKs - Microsoft Agent Framework
azure-ai-projects>=1.0.0b1
azure-ai-agents>=1.0.0b1
azure-identity>=1.19.0
azure-core>=1.30.0

# Azure Functions and Logic Apps
azure-functions>=1.20.0
azure-mgmt-logic>=10.0.0
azure-mgmt-web>=7.3.0

# AI and ML
openai>=1.54.0
python-dotenv>=1.0.0

# Development and Testing
pytest>=8.3.0
pytest-cov>=6.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0

# Code Quality
ruff>=0.8.0
black>=24.10.0
mypy>=1.13.0
types-requests>=2.32.0

# Utilities
pydantic>=2.10.0
requests>=2.32.0
aiohttp>=3.11.0
